backend/ms-gamification/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0.post1
SQLAlchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose==3.3.0
PyJWT==2.8.0
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
python-multipart==0.0.6
asyncpg==0.29.0
starlette==0.27.0
email-validator==2.1.0.post1 